requirements.txt
absl-py>=0.7.1
albumentations>=0.3.1
astor>=0.7.1
attrs>=19.1.0
backcall>=0.1.0
bleach>=3.1.0
certifi>=2019.6.16
cffi>=1.12.3
chardet>=3.0.4
cloudpickle>=1.0.0
cycler>=0.10.0
Cython>=0.29.13
cytoolz>=0.10.0
dask>=1.2.2
decorator>=4.4.0
defusedxml>=0.6.0
docopt>=0.6.2
entrypoints>=0.3
gast>=0.2.2
google-pasta>=0.1.7
grpcio>=1.16.1
h5py>=2.8.0
idna>=2.8
ijroi>=0.4.0
imageio>=2.5.0
imgaug>=0.2.9
jedi>=0.13.3
Jinja2>=2.10.1
joblib>=0.13.2
jsonschema>=3.0.2
Keras>=2.2.4
Keras-Applications>=1.0.7
Keras-Preprocessing>=1.0.9
kiwisolver>=1.1.0
Markdown>=3.1.1
MarkupSafe>=1.1.1
matplotlib>=3.1.1
mistune>=0.8.4
mock>=3.0.5
nbconvert>=5.5.0
nbformat>=4.4.0
networkx>=2.3
notebook>=6.0.0
numpy>=1.15.4
olefile>=0.46
pandas>=0.25.0
pandocfilters>=1.4.2
parso>=0.4.0
patsy>=0.5.1
pbr>=5.1.3
pexpect>=4.7.0
pickleshare>=0.7.5
Pillow>=6.1.0
PIMS>=0.4.1
pipreqs>=0.4.9
protobuf>=3.7.1
ptyprocess>=0.6.0
pycparser>=2.19
pydub>=0.23.1
Pygments>=2.4.0
pyparsing>=2.4.0
pyrsistent>=0.15.4
python-dateutil>=2.8.0
pytz>=2019.1
PyWavelets>=1.0.3
PyYAML>=5.1.2
pyzmq>=18.0.2
qtconsole>=4.4.4
read-roi>=1.5.2
requests>=2.22.0
scikit-image>=0.15.0
scikit-learn>=0.20.1
scipy>=1.1.0
seaborn>=0.9.0
Send2Trash>=1.5.0
six>=1.12.0
slicerator>=1.0.0
statsmodels>=0.10.1
tensorboard>=1.14.0
tensorflow-gpu>=2.1.0
termcolor>=1.1.0
terminado>=0.8.2
testpath>=0.4.2
tifffile>=0.15.1
toolz>=0.10.0
tornado>=6.0.3
traitlets>=4.3.2
urllib3>=1.25.3
wcwidth>=0.1.7
webencodings>=0.5.1
werkzeug>=0.15.3
widgetsnbextension>=3.5.1
wrapt>=1.11.2
yarg>=0.1.9
humanfriendly
toml
